requirements.txt
Flask
psycopg2-binary
pandas
python-dotenv
pytest
